pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crawler"
version = "0.0.1"
description = "Personal deep-monitoring crawler — sources to filtered signals via 8-stage pipeline."
requires-python = ">=3.12"
readme = "CONCEPT.md"
license = { text = "Proprietary" }
authors = [{ name = "Egor Okrepilov" }]

dependencies = [
    "pydantic>=2.5",
    "asyncpg>=0.29",
    "typing_extensions>=4.10",
    "praw>=7.7",
    "selectolax>=0.3",
    "langdetect>=1.0",
    "aiolimiter>=1.1",
    "tenacity>=8.0",
    "httpx>=0.27",
    "structlog>=24.0",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
crawler = "crawler.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
    "pyright>=1.1",
    "ruff>=0.4",
]

[tool.setuptools]
packages = [
    "crawler",
    "crawler.core",
    "crawler.storage",
    "crawler.plugins",
    "crawler.plugins.sources",
    "crawler.processing",
    "crawler.processing.stages",
    "crawler.api_core",
    "crawler.cli",
    "crawler.cli.commands",
]
include-package-data = true

[tool.setuptools.package-data]
"crawler.storage" = ["*.sql", "migrations/*.sql"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = false
check_untyped_defs = true
files = ["crawler"]

[tool.mypy-asyncpg]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
ignore = ["tests"]
